+++ src/test/java/com/googlecode/aviator/runtime/function/seq/set/SetContainsAnyFunctionTest.java
@@ -1,16 +1,15 @@
-package com.googlecode.aviator.runtime.function.seq.util;
+package com.googlecode.aviator.runtime.function.seq.set;
 
 import com.googlecode.aviator.AviatorEvaluator;
-import org.junit.Test;
+import junit.framework.TestCase;
 
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 
-public class SeqSubtractFunctionTest {
+public class SetContainsAnyFunctionTest extends TestCase {
 
-    @Test
-    public void call() {
+    public void testCall() {
         HashSet set1 =new HashSet();
         set1.add(1);
         set1.add(2);
@@ -23,7 +22,7 @@ public class SeqSubtractFunctionTest {
         Map env = new HashMap();
         env.put("set1",set1);
         env.put("set2",set2);
-        Object execute = AviatorEvaluator.execute("seqUtil.subtract(set2,set1)", env);
+        Object execute = AviatorEvaluator.execute("set.union(set1,set2)", env);
         System.out.println(execute);
     }
 }
