+++ src/test/java/com/googlecode/aviator/runtime/function/seq/set/SetDisjunctionFunctionTest.java
@@ -1,13 +1,15 @@
-package com.googlecode.aviator.runtime.function.seq.util;
+package com.googlecode.aviator.runtime.function.seq.set;
 
 import com.googlecode.aviator.AviatorEvaluator;
+import org.apache.commons.collections.CollectionUtils;
 import org.junit.Test;
 
+import java.util.Collection;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 
-public class SeqIntersectionFunctionTest {
+public class SetDisjunctionFunctionTest {
 
     @Test
     public void call() {
@@ -23,7 +25,9 @@ public class SeqIntersectionFunctionTest {
         Map env = new HashMap();
         env.put("set1",set1);
         env.put("set2",set2);
-        Object execute = AviatorEvaluator.execute("seqUtil.intersection(set1,set2)", env);
+        Collection union = CollectionUtils.union(set1, set2);
+        System.out.println(union);
+        Object execute = AviatorEvaluator.execute("set.disjunction(set1,set2)", env);
         System.out.println(execute);
     }
 }
