+++ src/test/java/com/googlecode/aviator/runtime/function/seq/set/SetIntersectionFunctionTest.java
@@ -1,15 +1,16 @@
-package com.googlecode.aviator.runtime.function.seq.util;
+package com.googlecode.aviator.runtime.function.seq.set;
 
 import com.googlecode.aviator.AviatorEvaluator;
-import junit.framework.TestCase;
+import org.junit.Test;
 
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 
-public class SeqContainsAnyFunctionTest extends TestCase {
+public class SetIntersectionFunctionTest {
 
-    public void testCall() {
+    @Test
+    public void call() {
         HashSet set1 =new HashSet();
         set1.add(1);
         set1.add(2);
@@ -22,6 +23,7 @@ public class SeqContainsAnyFunctionTest extends TestCase {
         Map env = new HashMap();
         env.put("set1",set1);
         env.put("set2",set2);
-        Object execute = AviatorEvaluator.execute("seqUtil.union(set1,set2)", env);
+        Object execute = AviatorEvaluator.execute("set.intersection(set1,set2)", env);
+        System.out.println(execute);
     }
 }
