+++ src/test/java/com/googlecode/aviator/runtime/function/seq/set/SetIsEqualFunctionTest.java
@@ -1,29 +1,28 @@
-package com.googlecode.aviator.runtime.function.seq.util;
+package com.googlecode.aviator.runtime.function.seq.set;
 
 import com.googlecode.aviator.AviatorEvaluator;
-import org.junit.Test;
+import junit.framework.TestCase;
 
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 
-public class SeqUnionFunctionTest {
+public class SetIsEqualFunctionTest extends TestCase {
 
-    @Test
-    public void call() {
+    public void testCall() {
         HashSet set1 =new HashSet();
         set1.add(1);
         set1.add(2);
         set1.add(3);
         HashSet set2 =new HashSet();
-        set2.add(5);
         set2.add(3);
-        set2.add(4);
+        set2.add(2);
+        set2.add(1);
 
         Map env = new HashMap();
         env.put("set1",set1);
         env.put("set2",set2);
-        Object execute = AviatorEvaluator.execute("seqUtil.union(set1,set2)", env);
+        Object execute = AviatorEvaluator.execute("set.intersection(set1,set2)", env);
         System.out.println(execute);
     }
 }
