+++ src/test/java/com/googlecode/aviator/runtime/function/seq/set/SetIsSubSeqFunctionTest.java
@@ -1,4 +1,4 @@
-package com.googlecode.aviator.runtime.function.seq.util;
+package com.googlecode.aviator.runtime.function.seq.set;
 
 import com.googlecode.aviator.AviatorEvaluator;
 import junit.framework.TestCase;
@@ -7,7 +7,7 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 
-public class SeqIsSubSeqFunctionTest extends TestCase {
+public class SetIsSubSeqFunctionTest extends TestCase {
 
     public void testCall() {
         HashSet set1 = new HashSet();
@@ -22,7 +22,7 @@ public class SeqIsSubSeqFunctionTest extends TestCase {
         Map env = new HashMap();
         env.put("set1", set1);
         env.put("set2", set2);
-        Object execute = AviatorEvaluator.execute("seqUtil.isSubSeq(set1,set2)", env);
+        Object execute = AviatorEvaluator.execute("set.isSubSeq(set1,set2)", env);
         System.out.println(execute);
     }
 }
